Splitting the Score into two quality models (resource consumption and performance)
